Fixing bug the the _LMFILES_ environment variable.  As the _LMFILES_ shrink in size the extra _LMFILES%03d_ were not being unset.  Made sure extra _LMFILES%03d_ are removed each time _LMFILES_ is set.
Added initial support for scheme (scm, guile).

Made output of ksh (bash, zsh) aliases recognize the escape charater ("\").
This was important to be able to escape "$".  Anytime an alias had a "$"
in it a function was always output.

Fixed a bug in search of the LOADEDMODULES environment variable.  Would not
correctly find a module if only a partial module name was given.  For
example "maya/3.0" would be found but "maya" would not.  In the process
of fixing this bug a possible memory corruption was also fixed.