Modified module switch so it unloaded last module first down to the module its swapping out and then reloaded everything in order.
Before it was unloading them starting at the module to be swapped out to
the end and then loading them in the same order.  Probably not a big deal
but code is actually shorter now and is probably more correct.

I also added -a and -append as alternatives to --append and --prepend for
module use.  (There were comments saying it would be nice to have that)

Kent