Create source docs for man pages.
This is still a work in progress, but I've converted the current
man pages over to pod (perl documentation format)
The idea being that it's easier to edit and can be converted to
various formats.  The manpages are very out of date, but this at least
will encourage me to keep them upto date and make it much easier to
work with.