Implement unload of x-resources and fix load
Implement unload action of x-resources defined for all shells. Unload of
x-resources is done by clearing string resource values with "xrdb
-merge" since there is no xrdb command that can clear one property (xrdb
-remove clears everything). For resources loaded from file, file content
is parsed with "xrdb -load" and each property it defines is also
cleared.

This commit also corrects the way x-resource value is parsed. Sometimes
x-resource value may be provided within resource name as the "x-resource
{Ileaf.popup.saveUnder: True}" example provided in manpage. So if no
value is provided, resource name is looked at to check if a value is
provided within string (space or colon are looked at).

Load of x-resources is also fixed by this commit for sh, csh and tcl to
handle resource name and value containing wildcard characters and fixed
for sh and csh to output well-behaving code to execute ("echo | xrdb
-merge" instead of "xrdb -merge <<EOF")