Save and restore collection specified as file path
Until now, it was only possible to save and restore a collection located
in default collection directory (under $HOME/.module). With this commit,
a file path can be specified to the save and restore commands.

On the contrary, saverm command has been enforced to use a name to target
the collection to remove in order to avoid this command to be able to
delete any file.

61-coll testsuite is adapted to test this new use case.