Refactor collection code a little bit
Introduce a getCollectionFilename procedure to query collection filename
and simplify error management code in cmdModuleSave and cmdModuleRestore
to improve code readability.

Also set separate procedures to generate the content of a collection
(formatCollectionContent procedure) and to read a collection file
(readCollectionContent procedure) in order to prepare further
enhancements on module collections.