Introduce collection target concept
Collection directory may sometimes be shared on multiple machines which
may use different modules setup. For instance modules users may access
with the same HOME directory multiple systems using different OS
versions. When it happens a collection made on machine 1 may be erroneous
on machine 2.

To handle this use case, we introduce here the concept of target for
module collections. By setting the MODULES_COLLECTION_TARGET environment
variable to something that can make a distinction between machines that
are incompatible with each other, only collections matching this target
will be valid on it.

When a target is set, only the collections made for that target will be
available to restore, list, show, rm, and saving will register the
target footprint (collection filename will be suffixed with
".$env(MODULES_COLLECTION_TARGET)". Collection target is not involved
when collection is specified as file path on show, restore and save
commands.

For example, the MODULES_COLLECTION_TARGET variable may be set with
results from commands like lsb_release, hostname, dnsdomainname, etc.