Introduce getLoadedModuleList proc
Defines a getLoadedModuleList procedure that returns a list of the
loaded modules by parsing the LOADEDMODULES environment variable.

This new procedure enables to gather in a single location the code to
parse LOADEDMODULES, thus it removes the need for many procedures to
know about "env" and "g_def_separator" global variables.