Introduce getModulePathList proc
Defines a getModulePathList procedure that returns a list of the
module paths by parsing the MODULEPATH environment variable.

This new procedure enables to gather in a single location the code to
parse MODULEPATH, thus it removes the need for many procedures to
know about "env" and "g_def_separator" global variables.

getModulePathList procedure also enables with its "behavior" parameter
to directly exit on error when MODULEPATH is not set. By default, it
returns an empty list. The "exiterronundef" behavior allows to closely
match previous behavior, when $env(MODULEPATH) were accessed without
being defined.