Standardize reportWarning procedure
Adapt reportWarning procedure to make it call "report" rather than
re-implementing a "puts stderr" procedure. Also add a "WARNING: "
prefix to all messages put through reportWarning to make this procedure
different from the report procedure.

Take also this opportunity to standardize some warning messages and to
change some 'reportError "WARNING' calls into reportWarning calls where
it was not necessary to raise the error counter (as it is warnings not
errors).