Harmonize reportErrorAndExit messages
As done for the other report* procedures, here we harmonize some
reportErrorAndExit messages to something simpler and clearer.

There is no need to make reportErrorAndExit call to reportError instead
of using "error" build-in proc since error proc raise a message which is
caught and output by reportError in the end.