Harmonize usage of exit when error occurs
When error occurs and exit is called, always quit with the same error
code: 1.

Add a call to "exit 1" when error is caught in main section when
processing module command.

Remove use of exit in cmdModuleInit procedure, replaced by call to
reportErrorAndExit, which will be caught at top level and go through the
"exit 1" added above.

End main section with an "exit 0" to explicit everything went well.