Merge main command handling in module proc
Make module command handling appear in a single location, the module
procedure, rather than having a duplication of code in the main section.

This modification simplifies the addition of new commands or the
adaptation of existing commands as everything has to be done only once.

Now calling module from a modulefile or a modulerc benefits from the
same command shortcuts than when it is called from the command line.

Moreover do not print anymore the module usage message when an invalid
command is typed, as this large message often hide the error printed on
the first line. An advice to type "module --help" is printed instead.