Put all command-line parsing under catch protection
Make command-line option parsing benefit from the catch (+reportError
+exit 1) applied to command parsing and running. This way it produces
same kind of error behavior when an invalid option is typed as when an
invalid command is typed.

Also put shell name parsing first in order to be able to renderError
appropriately when an invalid option is typed. Elsewhere the puts of
/bin/false on stdout was not made as no shell were set.

An invalid option containing only one '-' as first character will now
also produces an error. Before that only invalid options starting with
'--' were treated as error.

Improve testsuite on option treatment. Test long and short versions of
these options and test the returned output.