Merge list printing codes under displayElementList
Set a common procedure to manage the display of list of elements:
displayElementList. Lists can be displayed in a column fashion or with
only one element per line. Elements displayed may be prefixed by their
number. Also a header can be printed on top of the list.

Adapt cmdModuleList, cmdModuleSavelist, cmdModuleSearch, cmdModuleAvail
and cmdModuleAliases to use this new procedure.

With a single set of code to manage this aspect it will be easier to
improve it and all commands calling it will automatically benefit from
these improvements.