Improve column-mode in displayElementList
Change way to compute rows*cols display grid in order to maximize the
number of column to display per line. To do so a width is computed for
each column instead of having each column width based on the largest
one.

The new code takes a bit more time to proceed (as there is more stuff to
compute) but the additionnal time taken is not really discernable from
the user perspective: on a 'module avail' containing ~10 paths and ~400
modulefiles, there is less than 0.1 second extra time.

As display is now denser, two spaces are now used instead of one to
separate elements.

Thanks to Aurelien Degremont for crafting the right algorithm for this.