Check each command receives the correct number of arg
Improve module procedure to check for each module command that the
correct number of argument are provided. If an unexpected number of
argument is provided an error is raised as when an invalid command name
is specified. This way module exit with an error code and an error is
rendered in order to be checked from the shell.