testsuite: import and adapt 20-locate tests
Import the 20-locate test directory in use in Modules-C testsuite, that
was previously put aside for Modules-TCL. Adapt syntax of tests and
results to fit the output of Modules-TCL which sometimes slightly differ
from the C version.

At this stage, most of the tests from this 20-locate directory are
failling as aliases and versions support in Modules-TCL is not as
complete as it is in C version. Next commits will close the gap between
the two Modules flavors.