Fix module-info alias resolution
Call for resolveModuleVersionOrAlias procedure to deeply resolve name
provided by recursively looking at alias structures.

Improvement is made here to the resolveModuleVersionOrAlias procedure in
order to be able to only look for alias (and not alias and version) when
attempting resolution.

A g_aliasHash array is introduced here to be able to reversely look at
alias definition as Modules-C does. module-alias procedure has been
adapted to fill this new structure and resolveModuleVersionOrAlias makes
use of it if alias is searched and not found in g_moduleAlias array.