Register alias and version by the short module name
Store aliases and versions by the short module name rather than full
module path to get closer to the Modules-C behavior on alias and version
management. As a consequence an alias defined in a module path can apply
to a module stored in another module path. But this behavior is required
to make alias defined in global modulerc functional.

A getModuleNameVersion procedure is introduced to determine with a name
provided as argument the corresponding module name, version and
name/version. Module name is guessed from current module name when
shorthand version notation is used. Both name and version are guessed
from current module if name provided is empty.