Improve .modulerc/.version handling in getPathToModule
Adapt getPathToModule to parse .modulerc file at each level of the
specified module in order to be able to resolve any alias that may be
defined within the module directories. Means if module name
"mod/sub1/sub2" is passed to getPathToModule, the procedure will parse
any .modulerc file in "mod", "mod/sub1" and "mod/sub1/sub2" directories
(if these dirs exist) then it will attempt to resolve module name with
aliases and version found.

To behave like Modules-C, if both a .modulerc and a .version file exist
for the module name passed to getPathToModule, .modulerc is parsed then
.version. In this situation a default value set from .version will win
over a default value set in .modulerc. Same kind of change has been
applied to the listModules procedure.

A new argument named "search" has been added to the listModules
procedure to be able when called by getPathToModule with a "no_depth"
value to only get the content of a directory and not its sub elements
as sub-directory resolution, .modulerc and .version are handled within
getPathToModule.