Remove useless alias/version resolution in module proc
Aliases and versions are solved when getPathToModule proc is called and
getPathToModule is called by all command proc that need to resolve an
alias or a version.

So alias and version resolution is not required at the beginning of the
module procedure. Resolving any argument to the module procedure as
alias or version was also potentially erroneous.