Improve resolveModuleVersionOrAlias proc to detect loop
Adapt resolveModuleVersionOrAlias procedure to resolve only one name at
a time and pass a list of already resolved symbols for the current
resolution in order to detect infinite resolution loop.

When infinite resolution loop is detected error message is raised and
returned value is set to "*undef*" to behave like Modules C-version.