Fix module-info symbols resolution
Return all symbols defined by calling the getVersAliasList procedure.

This procedure has been enhanced to first attempt an alias resolution on
the name passed to it. Then each time a version or a default is found a
recursive call is made on this new name to get all versions or default
linked to the initial name.

getVersAliasList has also been enhanced to pass and get a list of
already resolved symbols for the current resolution in order to detect
infinite resolution loop.

Results returned by getVersAliasList are now sorted using the
dictionnary-style sort as already performed by listModules.

All these improvements made to the getVersAliasList will benefit to the
listModules procedure so "avail" command will return more accurate
versions and default information.