Include aliases as part of listModules results
Adapt listModules procedure to include to the results returned all the
module aliases defined through the analyzed directory. By doing so a
module avail or search commands automatically return all the aliases
defined in addition to regular modulefiles to get a complete view of
what is available.

In module-alias procedure the full path name of the file that declared
an alias is registered in a g_sourceAlias array. In listModules this
array is looked at to distinguish the aliases coming from the currently
analyzed directory.

Aliases are displayed from the module path where they are defined. A "@"
symbol is appended to the alias name, in the same way module-version are
displayed, in order to be able to distinguish an alias from a regular
modulefile. "@" has been retained as symbol to qualify module aliases as
this symbol is used by "ls" command to qualify symbolic links.

When --long display format is asked, the target of the alias is printed
after the alias name. No last modification time is displayed as it has
no meaning for a module alias.