Handle empty or separator path on add-path/unload-path
Improve add-path and unload-path to handle empty path {}, that can be
added or removed from a path list like on C-version.

Also if provided path is equal to the separator string, like ":", an
error is raised as this kind of path cannot be handled in path list
variables. This behavior here is different than on C-version where path
equals to ":" can be added or removed to path list variables.

Import or adapt add-path and unload-path relative tests (04*-append,
05*-prepend and 06*-remove) in 50-cmds suite.