Clarify module-info 'flags' and 'user' options
'flags' and 'user' are C-version specific options and they are not
relevant for Tcl-version. So they are present on Tcl-version to avoid
breaking modulefile using them but always return false or an empty
string to reflect the fact that these options are unmanaged here.

Import and adapt info-user and info-flags non-regression tests in
50-cmds suite.