Adapt 'system' to behave like described on manpage
Previously the 'system' modulefile command was adding the specified
string to the list of things to interpret on stdout. This commit aligns
the behavior of the 'system' command on the one implemented on Modules
C-version which is documented on modulefile(4) man page.

Now 'system' passes string to the 'exec' Tcl procedure with stdout
redirected to stderr since stdout would be parsed by the evaluating
shell. The exit status of the executed command is returned.

Adapt '140-system' test in 50-cmds suite accordingly.