Add dummy module-{log,trace,user,verbosity} procs
Add dummy procedures for the module-log, module-trace, module-user and
module-verbosity commands available on C-version. These commands do not
seem relevant to be implemented on Tcl-version of module, but having
dummy procedure for them enable support for modulefiles using these
commands. Without dummy procedures these modulefiles raise error as the
commands are not known.

The 4 dummy procedures report a warning message to let the user know
that even if the modulefile has been loaded/unloaded/etc, the
module-log, module-trace, module-user or module-verbosity calls were not
interpreted.

Import and adapt 100-loglevel, 101-badfac, 110-verbose and
115-verbose-msg tests in 50-cmds suite.