Add 'refresh' command as alias on 'reload' command
Add a 'refresh' command as available on C-version but instead of
reloading the volatile components (ie. shell aliases) only, here we
unload then reload all the currently loaded modules. This behavior is
different than on C-version but it ends with same results : shell
aliases are refreshed in sub-shells.

Import and adapt 07-refresh tests in 50-cmds suite.