Add support for 'exit' command in modulefile
Add support for use of 'exit' command within modulefile as described in
modulefile(4) man page. This command will cause the immediate cessation
of this module and any additional ones on the command line. This module
and the subsequent modules will not be listed as loaded. No environment
commands will be performed in the current module.

To achieve that a 'exitModfileCmd' procedure is added to superseed the
'exit' command during modulefile interpretation. When called this
procedure sets a flag that inhibit further modulefile interpretation.

Import and adapt 210-exit tests in 50-cmds suite.