Make 'switch' command consistent
Switch command were previsouly setting 'new' as 'old' and 'old' as 'new'
when 'new' is detected loaded. This behavior led to inconsistency
especially when the same 'module switch old new' command is called
twice. First time 'old' is loaded and 'new' is unloaded. Then for the
second call 'old' is unloaded and 'new' is loaded, and so on. In this
case, the result of the switch command depends on the current situation
which confuses most persons using it.

This commit makes the switch command idempotent by always ending up with
'old' unloaded and 'new' loaded, whatever the starting situation is.