Make 'switch' command handle a single argument
On a 'module switch [mod1] mod2', If 'mod1' is not specified, then it is
assumed to be the currently loaded module with the same root name as
'mod2'.

Introduce a 'getLoadedWithClosestName' procedure to look at currently
loaded modules and return the one with the closest name to the name
passed as argument.

Import and adapt 170-swap, 172-swap-app-delim, 173-swap-app-delim,
175-swap2, 177-swap-prep-delim and 178-swap-prep-delim tests to the
50-cmds suite.