Escape ':' in modshare var when used as path element in var
Some list variable may contain element with ':' in it, like for URL,
separated by a character different than ':'. However corresponding
modshare variable always use colon character as separator so we need to
escape colon used as variable content in modshare variable to avoid
issue.

To achieve that psplit and pjoin procedures are introduced as well as a
charUnescaped procedure, which is the reverse of the already existing
charEscaped procedure. With these new procedures path-like variable
escape or unescape $g_def_separator character used in element of the
variable when splitting or joining the whole variable.

Adapt 042-append-delim, 047-app-del-undo, 052-prepend-delim and
057-pre-del-undo tests in 50-cmds suite to check colon character is
correctly escaped in modshare variable.