Do not always treat 'default' version-name as a symbol
Do not always treat 'default' version-name as a symbol where
'module/default' is equal to 'module'. Sometimes 'module/default' can be
a real modulefile thus it should be treated like that.

To do so a 'default_is_special' argument has been added to the
'getModuleNameVersion' procedure in order to treat 'default'
as a special version-name only when working with 'module-version',
'module-info version' and 'module-info symbols' commands.

Add 035-defname tests to 50-cmds suite to check this behavior.