Add arg to test current {shell,shelltype}' in module-info
Add the possibility to pass a string to 'module-info shell' and
'module-info shelltype' to get a boolean test on currently used shell or
shelltype. This kind of behavior is already available on 'module-info
mode' to test module current mode.

These features are supported by C-version even if not documented.

modulefile documentation and 087-info-shells-exp tests in 50-cmds suite
have been adapted accordingly.