Handle multiple lines of module-whatis for same modfile
Treat g_whatis as a list where each module-whatis text-line is appended
to the previous lines.

In case of search, these lines are displayed only if they match the
search pattern (to get same behavior as on C-version)

As a side effect, if no 'module-whatis' directive are found in a
modulefile, this modulefile is never displayed on a 'module whatis' call
(same behavior as on C-version).

Adapt 210-exit whatis test in 50-cmds suite.