No exit raise on modfile error for avail/whatis cmds
Do not exit on an error raised by one modulerc or one modulefile for
global commands avail, whatis or search that parse the content of all
modulefile/modulerc found in the defined module paths.

By doing so results from these global commands will not be harmed if
error exists in a few modulefiles. The content of these modulefiles
will be ignored after the erroneous line. So in a whatis/search case if
the module-whatis definition is set before the erroneous line, the
modulefile will be taken into account. If the module-whatis definition
is set after the erroneous line, the modulefile will be ignored.

To achieve this goal a 'exit_on_error' argument has been added to the
'execute-modulerc' and 'execute-modulefile' procedures. By default this
argument is enabled but calls made from 'listModules' and
'cmdModuleSearch' disable it.

Add non-regression test to 055-whatis and 060-apropos in 70-maint suite
and to 030-multiple in 90-avail suite.