Refactor exit and ModulesVersion set of execute-modulerc
Refactor the execute-modulerc procedure to handle exit in case of
critical error as it is done on execute-modulefile procedure with the
slave interp evaluation returning an errorVal. Then this errorVal is
treated to check if a critical error as been encountered (errorVal=2).

To achieve this goal the way the ModulesVersion variable is set need to
be changed as the slave interp now returns errorVal not ModulesVersion.
A setModulesVersion procedure is added to enable the slave interp to
pass the value of the ModulesVersion variable in the master interp
context.