Non-zero exit code when exiting from execute-mod procs
When a critical error is encountered when interpreting a modulefile or a
modulerc from execute-modulefile or execute-modulerc procedures, exit as
done elsewhere. Means call for error code rendering and raise error to
top level, where exit command is called.

A new procedure called 'exitOnError' is added and it is called by
execute-modulefile and execute-modulerc procedures. Top level error
management code now does not always print an error message as in this
situation error message has already been displayed when the
reportInternalBug procedure was called.

Add non-regression tests to 190-load in 50-cmds suite and to 070-display
and 080-help in 70-maint suite.