Add module-info mode check against 'remove' and switch' value
For compatibility with C-version, enable to check current 'module-info
mode' against the 'remove' and 'switch' value. To achieve this change
'remove' is considered an alias of the 'unload' value. 'switch' is
checked against the current 'module-info command'.

084-info-mode-exp tests of 50-cmds suite have been adapted. In addition
020-unload and 022-unload2 tests of 95-version suite have also been
adapted as the modulefiles they check are using a 'module-info mode
remove' test.