www: Change website look and feel with Bootstrap
Adapt web pages to make use of the Bootstrap toolkit in order to get a
modern and easy to maintain web style and content. Bootstrap also helps
to get a responsive design so the website will adapt to the terminal
used to read it.

Thanks to Aurelien Degremont for pointing me to the right tool for this
task.