www: include NEWS files on website
Include the NEWS file of C and Tcl versions to the documentation portal
of Environment Modules.

Move C documentation to a 'c' directory to better distinguish from the
Tcl documentation that is stored in the 'tcl' directory. Hard links are
maintained on the Website not to break URL to man/module.html and
man/modulefile.html pages.