Fix MODULESHOME setup in autoinit
Set modulecmd.tcl script directory absolute then guess MODULESHOME from
it. Adapt this setup if modulecmd.tcl is located in a 'bin' directory
as in this case MODULESHOME is set to the upper directory (as 'init'
directory will be located at the same level than the 'bin' dir not
inside the 'bin' dir).

By the way, fix perl autoinit code to simply use 'argv0' rather than
'MODULESHOME' in order to set the module function as for the other
languages.