init: generate modulerc and .modulespath configs
Add makefile rules to generate the modulerc and .modulespath
configuration files. modulerc is generated in any case. By default it
hosts the MODULEPATH and LOADEDMODULES definition passed to the Makefile
with the 'modulepath' and 'loadedmodules' argument.

The 'setdotmodulespath' Makefile argument is also introduced. When set
to 'y', the .modulespath is generated and hosts the MODULEPATH
definition. In this case, modulerc config file only gets the
LOADEDMODULES definition or is set empty if no value is passed to the
'loadedmodules' argument.