Correct alias/version resolution on avail command
List of version symbols was erroneous on an avail command in case a
modulefile held symbols like default and was targeted by aliases. Avail
output was showing the aliases holding the symbols instead of the
modulefile.

getVersAliasList procedure was fixed to perform the alias resolution
attempt in a second phase of the procedure, to let chance to retrieve
symbols for the initial module name passed to the procedure.

030-multiple, 040-wild and 050-long tests of the 90-avail suite has been
enhanced with new tests to check this behavior. loc_rc6 modulefile has
been added to the testsuite modulefiles to reproduce this use case.

Thanks to Bert Wesarg and Levi Morrison for reporting this issue.