Fix getModuleNameVersion to translate mod name on demand only
Translate module name to currently interpreted module name when name
correspond to the last part this interpreted module only if translation
is enabled (was previously done in all cases). Translation is enabled
when usage of symbolic modulefile names are allowed, which means it is
enabled from 'module-version', 'module-alias' and 'module-info version'
and 'module-info symbols'.

Add non-reg tests in 020-load and 030-display of 80-deep suite to check
that.