Optimize grid size computation in displayElementList
Change algorithm strategy to compute the optimized number of column that
can be used to display in column-fashion a list of elements. Previous
algorithm was starting with a number of column equals to the number of
element in list. New algorithm starts closer to the solution with a
number of column equals to the basic solution where each column width is
the same (equals to the length of biggest element in list). This
approach reduces the number of iteration needed to reach best solution.

Also when a column width is changed currently computed row widths is
adapted only for this column and not entirely reset.

All these optimizations drastically improve grid size computation time.
New algorithm is approximatelly 12 time faster than the previous one.
There is no more an overhead for this display mode.