Separate file search from display treatment in listModules
Introduce findModules and getModules procedures to handle file search
for modulefiles and rc files interpretation. findModules is in charge of
searching paths for module-related files. getModules calls findModules
to get file list, then interprets any rc files found to consolidate a
list of modules to return.

listModules procedure now calls for getModules to get an array of
modules corresponding to the directory passed as argument. So focus of
listModules is now to prepare results for display.

This new methodolidy also corrects way directory where handled in case
of '--default' and '--latest' listing. On a '--default' listing, a
modulefile does not appear anymore if a dir is set default at the same
level. On a '--latest' listing: a dir does not appear anymore if set
default but not the latest.

Non-regression tests have been added in 020-avail of 91-sort suite to
ensure a correct behavior will be maintained over time.