Plug getPathToModule on findModules/getModules
Make getPathToModule procedure relying on findModules and getModules
procedures for module search and rc file interpretation, as listModules
now does. findModules and getModules have been adapted to also handle
getPathToModule use-cases.

Code to look for modules is now merged in a common set of procedures
which will help both search sides (getPathToModule/listModules) to get
immediate benefit from enhancement made on this code area.

As a side effect a 'module avail' on a module name which is an alias
will now correctly return this alias module. A test has been added in
020-single of 90-avail suite to check that.