Consider upper dir .version file in findModules
If searched mod is found next to a .version file (in same directory),
add this rc file to the list of those to interpret (like any modulerc
files found on the path between the modulepath dir and the dir of the
search mod).

With this fix 'module list' will correctly display the 'default' tag
along loaded modules when set via a .version file.

Adapt 045-listlong test in 70-maint suite to check this.