Add aliases to dir list of elements in getModules
When looking for an implicit default in a directory, now aliases are
part of the directory items so it can be returned as the last element in
it (highest numerically sorted version).

Add tests in 050-locrc of 20-locate suite to guaranty this behavior on
load command and tests in 020-avail of 91-sort suite to guaranty this
behavior on avail command.