Rework handling of error raised by checkValidModule
In checkValidModule procedure, register an entry in a g_invalid_mod_info
global array when a modulefile does not correctly set the magic cookie.
By doing so we are able to state during a modulefile search in
getPathToModule procedure that the modulefile was found but it was not
correctly defined.

This mechanism may be extended to other kind of modulefile error (like
access issues).

Add new non-regression tests to 190-load in 50-cmds suite for modulefile
specified as full path as new error handling better detect issues on
this use-case.