Move deep mod lookup logic from findModules to getModules
When a deep module is asked for searched to getModules (elt1/elt2/vers)
call findModules to search on upper element (elt1) in order to return
every modulerc or modulefiles related to this module.

Then getModules parses the array returned by findModules to build its
results with only the directories, modulefiles and aliases that
correspond to the asked search.

By doing so, some modulerc search logic can be eliminated from
findModules, as we will now always look at the whole deep module
content.