Change default search value of listModules to 'wild'
Previously set default value for 'search' argument was deprecated so
change it to 'wild' which will result when passed to getModules in the
addition of a wildcard to look at modules. This way the wildcard is only
added in conjunction of listModules searches as getPathToModule searches
should search for exact things.