Optimize getPathToModule when solved alias is upper-mod
Improve getPathToModule procedure to be able to solve any-kind of module
aliases within a single call rather than restarting new searches when
solved aliases are on the upper part (elt1/) of a current search
(elt1/elt2/) sharing same module root.

Add isSameModuleRoot procedure on the way to be able to check two
modules share the same base name.

Add non-regression tests in 020-load and 030-display of 80-deep suite to
guaranty that sub or upper aliases are correctly found.