Add symbolic version target to res list on getModules
When looking for modules, after looking at paths contents and module
aliases set within the current dir, look at symbolic versions set with
module-version to see if some match searched mod. If so, add the
modulefile target of these symbolic version to the result list if found
in path previously but not yet added to the result list.

With this new behavior a 'module avail mod/sym' will display the
relative 'mod/vers'. Previously nothing was returned.

Add non-regression tests in 020-single, 040-wild and 050-long of
90-avail suite to check this new behavior persist over time.