Disable modulerc command report on display-mode
Introduce a g_inhibit_dispreport global variable to be able to inhibit
the report of command set within a modulerc file when mode is set to
'display'.

This way a 'display' of a modulefile only outputs stuff from this
modulefile and not from the various .modulerc on the path toward this
modulefile.

Adapt most display-related non-regression tests to correctly check
returned output from start to end.